# schema/openapi.yaml
openapi: "3.0.3"
info:
  title: Slimeopolis API
  description: |
    Global Slime Emporium API — Cloudflare API Shield demonstration.
    Upload this schema to: Security → API Shield → Schema Validation.
  version: "1.0.0"
  contact:
    name: Slimeopolis API
    url: https://slimeopolis.example.com
servers:
  - url: https://slimeopolis.example.com/api
    description: Production (replace with your actual domain)

# ─── Security Schemes ───────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token from /auth/login or /auth/register.
        API Shield Session Identifier: configure the Authorization header
        as the session identifier for Volumetric Abuse Detection.
  schemas:
...
    CreateOrderRequest:
      type: object
      required:
        - shipping_name
        - shipping_address
        - shipping_city
        - shipping_country
        - shipping_postal_code
      properties:
        shipping_name:
          type: string
          minLength: 2
          maxLength: 200
          example: Jane Doe
        shipping_address:
          type: string
          minLength: 5
          maxLength: 500
          example: 123 Main Street
        shipping_city:
          type: string
          minLength: 2
          maxLength: 100
          example: New York
        shipping_country:
          type: string
          minLength: 2
          maxLength: 2
          pattern: "^[A-Z]{2}$"
          example: US
        shipping_postal_code:
          type: string
          minLength: 2
          maxLength: 20
          example: "10001"
        notes:
          type: string
          maxLength: 1000
          example: Please leave at door

    # ── Reviews ─────────────────────────────────────────────────────────────────
    CreateReviewRequest:
      type: object
      required: [rating, title, body]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        title:
          type: string
          minLength: 3
          maxLength: 120
          example: Absolutely love this slime!
        body:
          type: string
          minLength: 10
          maxLength: 2000
          example: The sakura scent is incredible. Ships fast, great quality.

    # ── User ────────────────────────────────────────────────────────────────────
...
paths:
...
  /products:
    get:
      operationId: listProducts
      summary: List all products with optional filters
      description: |
        API Shield: Sequence Analytics step 1 (browse).
        Schema Validation: query parameter type validation.
      tags: [Products]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 12
        - name: country
          in: query
          schema:
            type: string
            minLength: 2
            maxLength: 2
        - name: category
          in: query
          schema:
            type: string
            enum: [butter, cloud, clear, floam, crunchy, glossy, metallic, glitter, glow, clay]
        - name: featured
          in: query
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: Paginated product list
  /products/search:
    get:
      operationId: searchProducts
      summary: Full-text search across products
      description: |
        API Shield: PRIMARY Volumetric Abuse Detection target.
        Rate Limiting: 30 req/min per session.
        Schema Validation: q param is required string.
      tags: [Products]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 10
      responses:
        "200":
          description: Search results
        "422":
          description: q parameter required
  /products/{productId}:
    get:
      operationId: getProduct
      summary: Get a product by ID or slug
      description: |
        API Shield: Sequence Analytics step 2 (product detail view).
        Schema Validation: productId path parameter.
      tags: [Products]
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Product object with reviews
        "404":
          description: Product not found
  /products/{productId}/reviews:
    post:
      operationId: createReview
      summary: Post a review for a product
      description: |
        API Shield: Schema Validation — rating integer 1-5, title string, body string.
      tags: [Products]
      security:
        - BearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Review created
        "409":
          description: Already reviewed this product
        "422":
          description: Validation error

  # ── Cart ──────────────────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      summary: List current user's orders
      tags: [Orders]
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 10
      responses:
        "200":
          description: Paginated order list

    post:
      operationId: createOrder
      summary: Place an order from current cart
      description: |
        API Shield: FINAL Sequence Analytics step. Only reached legitimately after
        browse → product view → recommendations → add to cart. Bots that skip
        directly here are flagged by Sequence Analytics.
        Rate Limiting: 10 req/min per session.
        Schema Validation: all shipping fields validated.
      tags: [Orders]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order placed successfully
        "422":
          description: Validation error, empty cart, or insufficient stock
  /orders/{orderId}:
    get:
      operationId: getOrder
      summary: Get order details by ID
      tags: [Orders]
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order with line items
        "404":
          description: Order not found

  # ── Recommendations ───────────────────────────────────────────────────────────
...